test: adjust transaction builder tests
